Create config folder. Split webpack.js to webpack.dev.js, webpack.prod.js and webpack.common.js. Install webpack-merge and uglifyjs-webpack-plugin modules. Repoint start and build scripts to new config files path.
