+++ client/package.json
@@ -78,7 +78,7 @@
     "json-loader": "0.5.4",
     "node-sass": "^4.6.0",
     "object-assign": "^4.1.1",
-    "postcss-loader": "1.2.2",
+    "postcss-loader": "^2.0.9",
     "promise": "^8.0.1",
     "react-addons-test-utils": "^15.6.0",
     "react-dev-utils": "^0.5.2",
@@ -87,16 +87,17 @@
     "redux-mock-store": "^1.3.0",
     "sass-loader": "^6.0.6",
     "style-loader": "0.13.1",
+    "uglifyjs-webpack-plugin": "^1.1.0",
     "url-loader": "0.5.7",
     "webpack": "^3.8.1",
     "webpack-dev-server": "^2.9.4",
+    "webpack-merge": "^4.1.1",
     "whatwg-fetch": "^2.0.2"
   },
   "scripts": {
-    "build": "webpack",
-    "start": "webpack-dev-server --open",
+    "start": "webpack-dev-server --open --config ./config/webpack.dev.js",
+    "build": "webpack --config ./config/webpack.prod.js",
     "test": "jest",
     "test:watch": "npm test -- --watch"
-  },
-  "proxy": "http://localhost:8080"
+  }
 }
